Rename vmodels to mvvm
