+++ app/src/main/java/com/blockeq/stellarwallet/mvvm/exchange/ExchangeRepository.kt
@@ -1,4 +1,4 @@
-package com.blockeq.stellarwallet.vmodels.exchange
+package com.blockeq.stellarwallet.mvvm.exchange
 
 import android.app.Application
 import android.arch.lifecycle.LiveData
